src/third_party/blink/web_tests/http/tests/devtools/console/console-focus-expected.txt

Tests that interacting with the console gives appropriate focus.

Message count: 2

Running: testClickingWithSelectedTextShouldNotFocusPrompt
Prompt has focus: false
Viewport scrolled to top: true
Clicking message 0
Prompt has focus: false
Viewport scrolled to top: true

Running: testKeypressWithFocusedMessageShouldFocusPrompt
Prompt has focus: false
Viewport scrolled to top: true
Clicking message 0
Prompt has focus: true
Viewport scrolled to top: false

Running: testClickOutsideMessageListShouldFocusPromptAndMoveCaretToEnd
Prompt has focus: false
Viewport scrolled to top: true
Selection before: {"startLine":0,"startColumn":1,"endLine":0,"endColumn":1}
Clicking on container
Prompt has focus: true
Viewport scrolled to top: false
Selection after: {"startLine":0,"startColumn":6,"endLine":0,"endColumn":6}
